+++ CHANGELOG.md
@@ -7,4 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+* `LoadImage` trait
+* Implementation of those traits for the types in [`playdate-sys`](https://docs.rs/playdate-sys/0.2) behind the feature
+flag `playdate-sys-v02` (enabled by default).
+
+
 [Unreleased]: https://github.com/jcornaz/beancount_parser_2/compare/...HEAD